+++ app/src/main/java/eu/kanade/tachiyomi/data/track/simkl/dto/SimklSyncItem.kt
@@ -1,6 +1,7 @@
 package eu.kanade.tachiyomi.data.track.simkl.dto
 
 import eu.kanade.tachiyomi.data.database.models.anime.AnimeTrack
+import eu.kanade.tachiyomi.data.track.MalformedTrackerResponseException
 import eu.kanade.tachiyomi.data.track.TrackerManager
 import eu.kanade.tachiyomi.data.track.simkl.toTrackStatus
 import kotlinx.serialization.SerialName
@@ -8,9 +9,9 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class SimklSyncResult(
-    val anime: List<SimklSyncItem>?,
-    val tv: List<SimklSyncItem>?,
-    val movies: List<SimklSyncItem>?,
+    val anime: List<SimklSyncItem>? = null,
+    val tv: List<SimklSyncItem>? = null,
+    val movies: List<SimklSyncItem>? = null,
 ) {
     fun getFromType(type: String): List<SimklSyncItem>? {
         return when (type) {
@@ -24,14 +25,14 @@ data class SimklSyncResult(
 
 @Serializable
 data class SimklSyncItem(
-    val show: SimklSyncResultItem?,
-    val movie: SimklSyncResultItem?,
+    val show: SimklSyncResultItem? = null,
+    val movie: SimklSyncResultItem? = null,
     @SerialName("total_episodes_count")
-    val totalEpisodesCount: Long?,
+    val totalEpisodesCount: Long? = null,
     @SerialName("watched_episodes_count")
-    val watchedEpisodesCount: Double?,
+    val watchedEpisodesCount: Double? = null,
     @SerialName("user_rating")
-    val userRating: Int?,
+    val userRating: Int? = null,
 ) {
     fun toAnimeTrack(typeName: String, type: String, statusString: String): AnimeTrack {
         val resultData = getFromType(typeName)
@@ -40,8 +41,10 @@ data class SimklSyncItem(
             title = resultData.title
             remote_id = resultData.ids.simkl
             if (typeName != "movie") {
-                total_episodes = totalEpisodesCount!!
-                last_episode_seen = watchedEpisodesCount!!
+                total_episodes = totalEpisodesCount
+                    ?: throw MalformedTrackerResponseException("Simkl", "total episode count")
+                last_episode_seen = watchedEpisodesCount
+                    ?: throw MalformedTrackerResponseException("Simkl", "watched episode count")
             } else {
                 total_episodes = 1
                 last_episode_seen = if (statusString == "completed") 1.0 else 0.0
@@ -54,8 +57,8 @@ data class SimklSyncItem(
 
     fun getFromType(typeName: String): SimklSyncResultItem {
         return when (typeName) {
-            "show" -> show!!
-            "movie" -> movie!!
+            "show" -> show ?: throw MalformedTrackerResponseException("Simkl", "show")
+            "movie" -> movie ?: throw MalformedTrackerResponseException("Simkl", "movie")
             else -> throw Exception("Unknown type: $typeName")
         }
     }